+++ lib/generators/lite/encryption/templates/install.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 Lite::Encryption.configure do |config|
-  config.secret_key_base = ENV['SECRET_KEY_BASE']
-  config.secret_key_salt = ENV['SECRET_KEY_SALT']
+  config.encryption_iv = ENV['ENCRYPTION_IV']
+  config.encryption_key = ENV['ENCRYPTION_KEY']
+  config.encryption_salt = ENV['ENCRYPTION_SALT']
 end