+++ lib/lite/encryption/schemes/non_deterministic.rb
@@ -1,9 +1,5 @@
 # frozen_string_literal: true
 
-%w[key_generator message_encryptor message_verifier].each do |filename|
-  require "active_support/#{filename}"
-end
-
 module Lite
   module Encryption
     module Schemes
@@ -11,27 +7,21 @@ module Lite
 
         extend Lite::Encryption::Helpers::ClassMethods
 
-        KEY = ActiveSupport::KeyGenerator.new(
-          Lite::Encryption.configuration.secret_key_base
-        ).generate_key(
-          Lite::Encryption.configuration.secret_key_salt,
-          ActiveSupport::MessageEncryptor.key_len
-        ).freeze
-
-        private_constant :KEY
-
         def decrypt(value, opts = {})
-          encryptor.decrypt_and_verify(value, **opts)
+          cipher.decrypt_and_verify(value, **opts)
         end
 
         def encrypt(value, opts = {})
-          encryptor.encrypt_and_sign(value, **opts)
+          cipher.encrypt_and_sign(value, **opts)
         end
 
         private
 
-        def encryptor
-          @encryptor ||= ActiveSupport::MessageEncryptor.new(KEY)
+        def cipher
+          @cipher ||= ActiveSupport::MessageEncryptor.new(
+            Lite::Encryption.configuration.encryption_key,
+            cipher: Lite::Encryption::Key::CIPHER.name
+          )
         end
 
       end