+++ spec/lite/encryption/configuration_spec.rb
@@ -7,18 +7,18 @@ RSpec.describe Lite::Encryption::Configuration do
 
   describe '.configure' do
     it 'to be "foo"' do
-      Lite::Encryption.configuration.secret_key_base = 'foo'
+      Lite::Encryption.configuration.encryption_key = 'foo'
 
-      expect(Lite::Encryption.configuration.secret_key_base).to eq('foo')
+      expect(Lite::Encryption.configuration.encryption_key).to eq('foo')
     end
   end
 
   describe '.reset_configuration!' do
     it 'to be not "foo"' do
-      Lite::Encryption.configuration.secret_key_base = 'foo'
+      Lite::Encryption.configuration.encryption_key = 'foo'
       Lite::Encryption.reset_configuration!
 
-      expect(Lite::Encryption.configuration.secret_key_base).not_to eq('foo')
+      expect(Lite::Encryption.configuration.encryption_key).not_to eq('foo')
     end
   end