+++ spec/lite/encryption/key_spec.rb
@@ -4,15 +4,35 @@ require 'spec_helper'
 
 RSpec.describe Lite::Encryption::Key do
 
-  describe '#generate_base' do
-    it 'returns 128' do
-      expect(described_class.generate_base.size).to eq(128)
+  describe '#generate_iv' do
+    it 'returns 12' do
+      result = described_class.generate_iv
+
+      expect(result.size).to eq(12)
+    end
+  end
+
+  describe '#generate_key' do
+    it 'returns 32' do
+      result = described_class.generate_key
+
+      expect(result.size).to eq(32)
+    end
+  end
+
+  describe '#generate_password' do
+    it 'returns 32' do
+      result = described_class.generate_password
+
+      expect(result.size).to eq(32)
     end
   end
 
   describe '#generate_salt' do
     it 'returns 32' do
-      expect(described_class.generate_salt.size).to eq(32)
+      result = described_class.generate_salt
+
+      expect(result.size).to eq(32)
     end
   end