+++ spec/lite/encryption/schemes/deterministic_spec.rb
@@ -3,26 +3,20 @@
 require 'spec_helper'
 
 RSpec.describe Lite::Encryption::Schemes::Deterministic do
-  let(:decrypted_message) { 'test' }
-  let(:encrypted_message) { 'rW8QbA==' }
-
-  # before { Lite::Encryption.configuration.secret_key_base = CreditCard::DETERMINISTIC_KEY }
-  #
-  # after { Lite::Encryption.reset_configuration! }
 
   describe '.decrypt' do
-    it 'returns decrypted_message' do
-      decrypted_value = described_class.decrypt(encrypted_message)
+    it 'returns decrypted_email' do
+      decrypted_value = described_class.decrypt(encrypted_email)
 
-      expect(decrypted_value).to eq(decrypted_message)
+      expect(decrypted_value).to eq(decrypted_email)
     end
   end
 
   describe '.encrypt' do
-    it 'returns encrypted_message' do
-      encrypted_value = described_class.encrypt(decrypted_message)
+    it 'returns encrypted_email' do
+      encrypted_value = described_class.encrypt(decrypted_email)
 
-      expect(encrypted_value).to eq(encrypted_message)
+      expect(encrypted_value).to eq(encrypted_email)
     end
   end