+++ spec/lite/encryption/schemes/non_deterministic_spec.rb
@@ -3,20 +3,18 @@
 require 'spec_helper'
 
 RSpec.describe Lite::Encryption::Schemes::NonDeterministic do
-  let(:decrypted_message) { 'test' }
-
   describe '.decrypt' do
-    it 'returns message' do
-      encrypted_value = described_class.encrypt(decrypted_message)
+    it 'returns decrypted_email' do
+      encrypted_value = described_class.encrypt(decrypted_email)
       decrypted_value = described_class.decrypt(encrypted_value)
 
-      expect(decrypted_value).to eq(decrypted_message)
+      expect(decrypted_value).to eq(decrypted_email)
     end
   end
 
   describe '.encrypt' do
-    it 'returns true when encrypted message includes "=--"' do
-      encrypted_value = described_class.encrypt(decrypted_message)
+    it 'returns true when encrypted_value includes "=--"' do
+      encrypted_value = described_class.encrypt(decrypted_email)
 
       expect(encrypted_value.include?('=--')).to eq(true)
     end