+++ spec/spec_helper.rb
@@ -9,7 +9,7 @@ Lite::Encryption.reset_configuration!
 
 spec_path = Pathname.new(File.expand_path('../spec', File.dirname(__FILE__)))
 
-%w[config models].each do |dir|
+%w[config helpers models].each do |dir|
   Dir.each_child(spec_path.join("support/#{dir}")) do |f|
     load(spec_path.join("support/#{dir}/#{f}"))
   end
@@ -30,4 +30,6 @@ RSpec.configure do |config|
     temp_path = spec_path.join('generators/lite/tmp')
     FileUtils.remove_dir(temp_path) if File.directory?(temp_path)
   end
+
+  config.include ContextHelper
 end