+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.1.0] - 2021-04-12
 ### Added
 - Added deterministic encryption
+### Changed
+- Update configuration methods
 
 ## [1.0.1] - 2020-07-03
 ### Added